Setup
New structure of project. Setup.py added
